refactor(backend): rename package with handlers deps to utils
