+++ backend/src/common/utils/activation_code_generator.py
