+++ backend/src/users/service/event_handlers.py
@@ -1,6 +1,6 @@
 from src.common.adapters.activation_code_storage import ActivationCodeStorage
-from src.common.dependencies.activation_code_generator import ActivationCodeGenerator
-from src.common.dependencies.notificator import Notificator
+from src.common.utils.activation_code_generator import ActivationCodeGenerator
+from src.common.utils.notificator import Notificator
 from src.common.domain.events import DomainEvent
 from src.common.service.uow import UnitOfWork
 from src.users.domain.events import (
