+++ backend/src/users/service/handlers_utils.py
@@ -1,6 +1,6 @@
 from src.common.adapters.activation_code_storage import ActivationCodeStorage
-from src.common.dependencies.activation_code_generator import ActivationCodeGenerator
-from src.common.dependencies.notificator import Notificator
+from src.common.utils.activation_code_generator import ActivationCodeGenerator
+from src.common.utils.notificator import Notificator
 from src.common.service.exceptions import PasswordValidationError, UserNotFound
 from src.common.service.uow import UnitOfWork
 from src.users.domain.model import User
