+++ backend/tests/fakes.py
@@ -1,7 +1,7 @@
 from uuid import UUID
 
 from src.common.adapters.activation_code_storage import ActivationCodeStorage
-from src.common.dependencies.notificator import Notificator
+from src.common.utils.notificator import Notificator
 from src.common.service.exceptions import UserNotFound, WishlistNotFound
 from src.common.service.uow import UnitOfWork
 from src.users.adapters.user_repository import UserRepository
